requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
huggingface_hub>=0.24.0


